Standardize HttpClient registration pattern in Blazor Client
Move all wildlife-related HTTP client services (IMunicipalityService,
ISpeciesService, IExportService, IWildlifeService) from FaunaFinder.Client
to Wildlife.Application.Client project, matching the pattern used by
Identity.Application.Client.

- Add AddWildlifeClient(Uri baseAddress) extension method that registers
  all wildlife HTTP clients with proper base address configuration
- Merge IWildlifeClient/WildlifeClient with IWildlifeService/WildlifeApiService
- Update Program.cs to use single AddWildlifeClient() call instead of
  individual AddHttpClient<> registrations
- Add Pagination.Contracts reference to Wildlife.Application.Client
- Update _Imports.razor and page using statements for new namespace

This standardizes the client registration pattern across features and
reduces code duplication in the main Blazor client project.

Fixes #136
