+++ src/FaunaFinder.Client/Pages/Wildlife/MySightings.razor
@@ -1,6 +1,5 @@
 @page "/sightings"
 @using Microsoft.AspNetCore.Authorization
-@using FaunaFinder.Client.Services.Api
 @using FaunaFinder.Wildlife.Contracts
 @inject IWildlifeService WildlifeService
 @inject NavigationManager Navigation
