+++ src/FaunaFinder.Client/Pages/Wildlife/ReportSighting.razor
@@ -1,6 +1,5 @@
 @page "/sightings/report"
 @using Microsoft.AspNetCore.Authorization
-@using FaunaFinder.Client.Services.Api
 @using FaunaFinder.Wildlife.Contracts
 @inject IWildlifeService WildlifeService
 @inject NavigationManager Navigation
