+++ src/FaunaFinder.Client/_Imports.razor
@@ -11,6 +11,7 @@
 @using FaunaFinder.Client.Pages
 @using FaunaFinder.Client.Components
 @using FaunaFinder.Client.Services.Api
+@using FaunaFinder.Wildlife.Application.Client
 @using FaunaFinder.Client.Services.DarkMode
 @using FaunaFinder.Client.Services.Localization
 @using FaunaFinder.i18n.Contracts
