+++ src/Features/Wildlife/FaunaFinder.Wildlife.Application.Client/FaunaFinder.Wildlife.Application.Client.csproj
@@ -12,6 +12,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\FaunaFinder.Wildlife.Contracts\FaunaFinder.Wildlife.Contracts.csproj" />
+    <ProjectReference Include="..\..\..\Common\FaunaFinder.Pagination.Contracts\FaunaFinder.Pagination.Contracts.csproj" />
   </ItemGroup>
 
 </Project>
