+++ src/Features/Wildlife/FaunaFinder.Wildlife.Application.Client/IExportService.cs
@@ -1,4 +1,4 @@
-namespace FaunaFinder.Client.Services.Api;
+namespace FaunaFinder.Wildlife.Application.Client;
 
 public interface IExportService
 {
