+++ src/Features/Wildlife/FaunaFinder.Wildlife.Application.Client/SpeciesApiService.cs
@@ -3,7 +3,7 @@ using FaunaFinder.Wildlife.Contracts.Dtos;
 using FaunaFinder.Wildlife.Contracts.Parameters;
 using FaunaFinder.Pagination.Contracts;
 
-namespace FaunaFinder.Client.Services.Api;
+namespace FaunaFinder.Wildlife.Application.Client;
 
 public sealed class SpeciesApiService : ISpeciesService
 {
